pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "typr"
version = "0.1.0"
description = "Speech-to-text for Linux with OpenAI Whisper"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Murtaza" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "PyQt6>=6.5.0",
    "pyaudio>=0.2.14",
    "httpx>=0.25.0",
    "websockets>=12.0",
    "dbus-python>=1.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-qt>=4.2",
    "mypy>=1.0",
    "ruff>=0.1",
]

[project.scripts]
typr = "typr.__main__:main"

[project.gui-scripts]
typr-gui = "typr.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/typr"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP"]
